requirements.txt
anthropic
playwright
python-dotenv
pandas
requests
jinja2
mjml
pydantic
